project/requirements.txt
asgiref==3.2.3
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreschema==0.0.4
Django==3.0.3
djangorestframework==3.11.0
drf-yasg==1.20.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.1
packaging==21.3
pycodestyle==2.8.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2019.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.3.0
swagger-spec-validator==2.7.4
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.7.0
